Move 'albuquerque' fixture to conftest
I keep re-defining this fixture. Move it to conftest so it is
available wherever it is needed.
