+++ pvanalytics/tests/features/test_daytime.py
@@ -6,11 +6,6 @@ from pvlib.location import Location
 from pvanalytics.features import daytime
 
 
-@pytest.fixture(scope='module')
-def albuquerque():
-    return Location(35, -106, altitude=1500)
-
-
 @pytest.fixture(scope='module',
                 params=['H', '15T', pytest.param('T', marks=pytest.mark.slow)])
 def clearsky_january(request, albuquerque):
