+++ pvanalytics/tests/quality/test_weather.py
@@ -2,7 +2,6 @@
 import pytest
 import pandas as pd
 import numpy as np
-from pvlib import location
 from pandas.util.testing import assert_series_equal
 from pvanalytics.quality import weather
 
@@ -90,11 +89,8 @@ def test_wind_limits(weather_data):
     )
 
 
-def test_module_temperature():
+def test_module_temperature(albuquerque):
     """Module temperature is correlated with GHI."""
-    albuquerque = location.Location(
-        35.0844, -106.6504, altitude=5312, tz='MST'
-    )
     times = pd.date_range(
         start='01/01/2020',
         end='03/01/2020',
